requirements.txt
aiogram
python-dotenv
requests
gigachat